--- a/src/System.Private.DataContractSerialization/src/System/Runtime/Serialization/ClassDataContract.cs
+++ b/src/System.Private.DataContractSerialization/src/System/Runtime/Serialization/ClassDataContract.cs
@@ -303,12 +303,17 @@ namespace System.Runtime.Serialization
                     }
                     catch
                     {
-                        _makeNewInstance = () => null;
+                        _makeNewInstance = ReturnNull;
                     }
                 }
 
                 return _makeNewInstance;
             }
+        }
+
+        private object ReturnNull()
+        {
+            return null;
         }
 
         internal object CreateNewInstanceViaDefaultConstructor(ConstructorInfo ci)