… src/System.Runtime.Serialization.Xml/tests/DataContractSerializer.cs
… src/System.Runtime.Serialization.Xml/tests/Performance/PerformanceTestsCommon.cs
… src/System.Runtime.Serialization.Xml/tests/SerializationTypes.cs