The fix improves the performance of reflection based de-serialization. The basic idea was to cache useful information acquired via reflection when de-serializing collections and create delegates for reflection invokes. 

The table below shows the results of the perf test runs,

| DeSerializationTest    | numberOfRuns | testSize | Baseline | Fix      | Diff    |
|------------------------|--------------|----------|----------|----------|---------|
| ByteArray              | 100          | 1024     | 0.563    | 0.597    | 6.04%   |
| ByteArray              | 5            | 1048576  | 21.4     | 21.8     | 1.87%   |
| String                 | 10000        | 128      | 23.7     | 24.6     | 3.80%   |
| String                 | 10000        | 1024     | 48.6     | 47.9     | -1.44%  |
| ListOfInt              | 1000         | 128      | 611      | 64.9     | -89.38% |
| ListOfInt              | 100          | 1024     | 494      | 46.8     | -90.53% |
| ListOfInt              | 1            | 1048576  | 5.07E+03 | 474      | -90.65% |
| Dictionary             | 1000         | 128      | 694      | 497      | -28.39% |
| Dictionary             | 100          | 1024     | 539      | 368      | -31.73% |
| SimpleType             | 10           | 1        | 0.616    | 0.31     | -49.68% |
| SimpleType             | 1            | 15       | 2.21E+03 | 1.07E+03 | -51.58% |
| ISerializable          | 10000        | -1       | 73.2     | 76.2     | 4.10%   |
| XmlElement             | 10000        | -1       | 34.6     | 35.2     | 1.73%   |
| ArrayOfSimpleType      | 1000         | 128      | 1.90E+03 | 394      | -79.26% |
| ArrayOfSimpleType      | 100          | 1024     | 1.53E+03 | 294      | -80.78% |
| ListOfSimpleType       | 1000         | 128      | 1.77E+03 | 371      | -79.04% |
| ListOfSimpleType       | 100          | 1024     | 1.43E+03 | 290      | -79.72% |
| DictionaryOfSimpleType | 1000         | 128      | 1.81E+03 | 934      | -48.40% |
| DictionaryOfSimpleType | 100          | 1024     | 1.44E+03 | 721      | -49.93% |