@shmao 2016-06-22
I added more commits in this PR. The new change was optimization on de-serializing value types. I added a new perf test for that case. Please see below for the perf result.

| DeSerializationTest       | numberOfRuns | testSize | Baseline (ms) | Fix (ms)  | Diff    |
|---------------------------|--------------|----------|----------|------|---------|
| SimpleStructWithPropertie | 10000        | 1        | 80.8     | 63.1 | -21.91% |