@mconnew 2016-06-22
Minor nits, otherwise LGTM